requirements.txt
fastapi
uvicorn
python-multipart
PicImageSearch
Pillow
httpx
aiofiles
streamlit
requests
PyJWT
slowapi
celery[redis]
